Separate building from packaging for distribution
All stuff related to packaging of RingoJS for distribution is now
contained in the dist/ directory. The aims of this patch are threefold:

- Ease packaging from an unclean ringojs working directory, by
  explicitly listing the files to be packaged.

- Allow packaging of "development snapshots" (as tar/zipballs) from
  untagged commits by using git-describe to add an <offset>-<sha1>
  suffix to the created packages.

- Create all packages, including the .deb's, in a single directory.

Packaging-related targets (package, dpkg) have been removed from
build.xml and are now handled by a makefile. Files to be packaged for
distribution are explicitly listed in dist/filelist (which is used as
rsync filter rules merge-file).

Debian packaging scripts have been moved to dist/debian/. The only
changes in Debian scripts are to adapt debian/rules to the new directory
layout, and in debian/ringojs.install to explicitly list the apps to be
bundled (instead of just bundling the complete apps/ directory).
Finally, we no longer attempt to create and sign a Debian .changes file,
as we currently don't use it.